requirements.txt
langchain
langchain-community
langchain-core
ollama
openai-whisper
SpeechRecognition
sounddevice
PyAudio
kokoro-onnx
onnxruntime
torch
torchvision
pillow
mss
opencv-python
requests
websockets
